+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add `Tests` section into Kaoto view
   - list existing Citrus tests in a current workspace
   - create a new Citrus Test file button
+  - run single test or all tests in a folder
 
 # 2.9.0